exclusions option, default connections fixed

+++ Connections.cs
@@ -13,7 +13,6 @@ namespace ProfileBackupTool
 {
     public partial class Connections : Form
     {
-        string defaultServer = Properties.Settings.Default.DefaultServer;
 
         public Connections()
         {
@@ -27,18 +26,17 @@ namespace ProfileBackupTool
                     string line;
                     while ((line = r.ReadLine()) != null)
                     {
-                        lines.Add(line);
+                        if(line != "")
+                        {
+                            lines.Add(line);
+                        }
+
                     }
                 }
 
                 foreach (string line in lines)
                 {
                     ConnectionsList.Items.Add(line);
-
-                    if(line == defaultServer)
-                    {
-                        SetAsDefaultCheckbox.Checked = true;
-                    }
                 }
             }
             catch
@@ -59,6 +57,8 @@ namespace ProfileBackupTool
             {
                 ValidateAccesibility();
             }
+
+            // TODO: add server without verifying
         }
 
         public void ValidateAccesibility()
@@ -70,14 +70,25 @@ namespace ProfileBackupTool
             {
                 File.Open(ServerAddress + "dummy.txt", FileMode.Create).Close();
                 File.Delete(ServerAddress + "dummy.txt");
-                ConnectionsList.Items.Add(ServerAddress);
-                File.AppendAllText("config\\connections.txt", ServerAddress + Environment.NewLine);
- 
+
+                if (ServerAddress != "")
+                {
+
+                    if (SetAsDefaultCheckbox.Checked)
+                    {
+                        Properties.Settings.Default.DefaultServer = ServerAddress;
+                        Properties.Settings.Default.Save();
+                        SetAsDefaultCheckbox.Checked = true;
+                    }
+
+                    ConnectionsList.Items.Add(ServerAddress);
+                    File.AppendAllText("config\\connections.txt", ServerAddress + Environment.NewLine);
+                }
+
             }
-            catch (Exception e)
+            catch
             {
-                MessageBox.Show(e.ToString());
-             //   MessageBox.Show("Unable to contact server.");
+                MessageBox.Show("Unable to contact network resource. Make sure UNC is valid .");
             }
         }
 
@@ -96,40 +107,46 @@ namespace ProfileBackupTool
 
         private void RemoveConnectionButton_Click(object sender, EventArgs e)
         {
-            string item = ConnectionsList.SelectedItem.ToString();
-            string tempFile = Path.GetTempFileName();
-            string filePath = "config\\connections.txt";
-            try
+            if (ConnectionsList.Items.Count != 0)
             {
-                using (var sr = new StreamReader(filePath))
+                try
                 {
-                    using (var sw = new StreamWriter(tempFile))
+                    string item = ConnectionsList.SelectedItem.ToString();
+                    string tempFile = Path.GetTempFileName();
+                    string filePath = "config\\connections.txt";
+
+                    using (var sr = new StreamReader(filePath))
                     {
-                        string line;
-                        while ((line = sr.ReadLine()) != null)
+                        using (var sw = new StreamWriter(tempFile))
                         {
-                            if (line != item)
+                            string line;
+                            while ((line = sr.ReadLine()) != null)
                             {
-                                sw.WriteLine(line);
+                                if (line != item)
+                                {
+                                    sw.WriteLine(line);
+                                }
                             }
                         }
                     }
+
+                    File.Delete(filePath);
+                    File.Move(tempFile, filePath);
+                    ConnectionsList.Items.Remove(ConnectionsList.SelectedItem);
+                    SetAsDefaultCheckbox.Checked = false;
                 }
-            }
-            catch(Exception we)
-            {
-                MessageBox.Show(we.ToString());
-            }
-            File.Delete(filePath);
-            File.Move(tempFile, filePath);
-            ConnectionsList.Items.Remove(ConnectionsList.SelectedItem);
+                catch 
+                {
+               //     MessageBox.Show(we.ToString());
+                }
+            }       
         }
 
         private void ConnectionsList_SelectedIndexChanged(object sender, EventArgs e)
         {
             try
             {
-                if (ConnectionsList.SelectedItem.ToString() == defaultServer)
+                if (ConnectionsList.SelectedItem.ToString() == Properties.Settings.Default.DefaultServer)
                 {
                     SetAsDefaultCheckbox.Checked = true;
                 }
